installation process

+++ deploy/create_deploy_directory.php
@@ -10,14 +10,16 @@ function remove_directory($path) {
 			unlink($path."/".$filename);
 	}
 	closedir($dir);
-	rmdir($path);
+	if (!@rmdir($path))
+		rmdir($path);
 }
 
 if (file_exists($_POST["path"]))
 	remove_directory($_POST["path"]);
 if (file_exists($_POST["path"])) die("Unable to remove directory ".$_POST["path"]);
 
-if (!mkdir($_POST["path"])) die("Unable to create directory ".$_POST["path"]);
+if (!@mkdir($_POST["path"]))
+	if (!mkdir($_POST["path"])) die("Unable to create directory ".$_POST["path"]);
 if (!file_exists($_POST["path"])) die("Unable to create directory ".$_POST["path"]);
 
 ?>
